fix: various fixes
Fee estimation: now returns in gwei instead of wei as intended.

Account store: assign anonymous function to proxy to fix `this`
context (was not setting to plugin store). Also, set the object property
value on the set trap (was not persisting account changes to store).

Still a WIP, various TODOs throughout.